ref: ModalManager -> 모든 CRUD 요청에서 invalidQuries 적용
