+++ components/Folder/FolderTag.tsx
@@ -13,7 +13,7 @@ const FolderTag = ({
     "w-[100px] h-[35px] px-[15px] text-sm whitespace-nowrap truncate border border-purple100 rounded-md hover:bg-purple100 hover:text-white";
 
   const handleSubmit = (selectedFolderId: number | string) => {
-    setQueryKeys((prev: any) => ({ ...prev, folder: selectedFolderId }));
+    setQueryKeys((prev: any) => ({ ...prev, folderId: selectedFolderId }));
   };
 
   return (
